Create a new virtual environment 'venv2' and reinstall dependencies from dev.txt (which includes everything from common.txt).
The path to the 'venv', the original virtual environment folder,
had been changed when I shifted folders around in the project.
This apparently causes the venv folder to reference the system
packages rather than the project-specific packages in
'project_folder/api/venv/bin'. So I've been using the wrong
packages for a while...

.gitignore venv2

Build the frontend via the command 'npm run build'.
This creates a 'build' folder which is basically a truncated
version of the 'public' folder and all of the js, css and images.